gtk: Modularize GTK display
Allows GTK to be compiled as a module and loaded at runtime. If GTK is
not the display being used, the module will not be loaded. In the case
that the GTK module is not present, if GTK was specified explicitly by
the user, qemu will exit with an error, otherwise it will default to
display type None.

This should give improvements in terms of performance, since much less
libraries will be loaded at startup if GTK is not being used, and in
terms of packaging, since GTK should be able to be packaged separately
in binary package managers now.

Signed-off-by: Colin Lord <clord@redhat.com>